requirements.txt
django==4.2.4
psycopg2-binary>=2.9.9
django-environ
